Warning remove. delete old unused 68k files.

+++ library/argz_add.c
@@ -47,7 +47,7 @@
 error_t
 argz_add(char **argz, size_t *argz_len, const char *str)
 {
-    int len_to_add = 0;
+    size_t len_to_add = 0;
     size_t last = *argz_len;
 
     if (str == NULL)
